unittest move to selftests folder
