+++ selftests/azure/azure_locations_australiaeast.json
