+++ selftests/azure/azure_locations_brazilsouth.json
