+++ selftests/azure/azure_locations_eastus.json
