+++ selftests/azure/azure_locations_eastus2.json
