+++ selftests/azure/azure_locations_northeurope.json
