+++ selftests/azure/azure_locations_notreal.json
