+++ selftests/azure/azure_locations_southcentralus.json
