+++ selftests/azure/azure_locations_southeastasia.json
