+++ selftests/azure/azure_locations_uksouth.json
