+++ selftests/azure/azure_locations_westeurope.json
