+++ selftests/azure/azure_locations_westus2.json
