+++ selftests/combinators/test_batch_combinator.py
