+++ selftests/test_platform.py
@@ -17,9 +17,9 @@ from lisa.environment import (
 )
 from lisa.feature import Feature
 from lisa.platform_ import Platform, WaitMoreResourceError, load_platform
-from lisa.tests.test_environment import generate_runbook as generate_env_runbook
 from lisa.util import LisaException, constants, plugin_manager
 from lisa.util.logger import Logger
+from selftests.test_environment import generate_runbook as generate_env_runbook
 
 
 @dataclass
