+++ selftests/test_search_space.py
