+++ selftests/test_secret.py
