+++ selftests/test_variable.py
